Fill form with language features
selectionOptions.json is now also being generated
by jsonCreation.py. This file takes a look at all the phonological
processes of all the langs in the json/ folder and creates
td table compatible features that correspond to the field ID's
in the form we're using for searching. This way, selection options
on the phonemes page are always up to date with what actually is
seen in the language data.