Make the backend work for: finding languages that have processes that have all these characteristics.
When you process form characteristics: look through all the processes in a language. If there is a process in a language that satisfies all the filters: add that language to langset. Otherwise, langset remains empty.

**Narrative that we want to have**: look for languages that have nasalization processes that affect stops

In this commit, I expanded out natural classes so we can search on the individual segments using the IPA keyboard. Similarly, I contracted individual segments into natural classes so we can search on natural classes from drop-down. This is to resolve inconsistencies in the Excel spreadsheet coding where sometimes grad students used natural classes, sometimes not.